Extracted default dir for file cache test
