+++ src/test/java/net/snowflake/client/core/FileCacheManagerTest.java
@@ -24,9 +24,7 @@ import java.util.stream.Stream;
 import net.snowflake.client.annotations.RunOnLinuxOrMac;
 import net.snowflake.client.category.TestTags;
 import net.snowflake.client.jdbc.BaseJDBCTest;
-import net.snowflake.client.jdbc.SnowflakeUtil;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Tag;
@@ -142,105 +140,6 @@ class FileCacheManagerTest extends BaseJDBCTest {
         ex.getMessage()
             .contains(
                 "Unable to access the file/directory to check the permissions. Error: java.nio.file.NoSuchFileException:"));
-  }
-
-  @Test
-  public void shouldCreateCacheDirForLinuxXDG() {
-    try (MockedStatic<Constants> constantsMockedStatic = Mockito.mockStatic(Constants.class)) {
-      constantsMockedStatic.when(Constants::getOS).thenReturn(Constants.OS.LINUX);
-      try (MockedStatic<SnowflakeUtil> snowflakeUtilMockedStatic =
-          Mockito.mockStatic(SnowflakeUtil.class)) {
-        snowflakeUtilMockedStatic
-            .when(() -> SnowflakeUtil.systemGetEnv("XDG_CACHE_HOME"))
-            .thenReturn("/XDG/Cache/");
-        try (MockedStatic<FileUtil> fileUtilMockedStatic = Mockito.mockStatic(FileUtil.class)) {
-          fileUtilMockedStatic.when(() -> FileUtil.isWritable("/XDG/Cache/")).thenReturn(true);
-          File defaultCacheDir = FileCacheManager.getDefaultCacheDir();
-          Assertions.assertNotNull(defaultCacheDir);
-          Assertions.assertEquals("/XDG/Cache/snowflake", defaultCacheDir.getAbsolutePath());
-        }
-      }
-    }
-  }
-
-  @Test
-  public void shouldCreateCacheDirForLinuxWithoutXDG() {
-    try (MockedStatic<Constants> constantsMockedStatic = Mockito.mockStatic(Constants.class)) {
-      constantsMockedStatic.when(Constants::getOS).thenReturn(Constants.OS.LINUX);
-      try (MockedStatic<SnowflakeUtil> snowflakeUtilMockedStatic =
-          Mockito.mockStatic(SnowflakeUtil.class)) {
-        snowflakeUtilMockedStatic
-            .when(() -> SnowflakeUtil.systemGetEnv("XDG_CACHE_HOME"))
-            .thenReturn(null);
-        snowflakeUtilMockedStatic
-            .when(() -> SnowflakeUtil.systemGetProperty("user.home"))
-            .thenReturn("/User/Home");
-        try (MockedStatic<FileUtil> fileUtilMockedStatic = Mockito.mockStatic(FileUtil.class)) {
-          fileUtilMockedStatic.when(() -> FileUtil.isWritable("/User/Home")).thenReturn(true);
-          File defaultCacheDir = FileCacheManager.getDefaultCacheDir();
-          Assertions.assertNotNull(defaultCacheDir);
-          Assertions.assertEquals("/User/Home/.cache/snowflake", defaultCacheDir.getAbsolutePath());
-        }
-      }
-    }
-  }
-
-  @Test
-  public void shouldCreateCacheDirForWindows() {
-    try (MockedStatic<Constants> constantsMockedStatic = Mockito.mockStatic(Constants.class)) {
-      constantsMockedStatic.when(Constants::getOS).thenReturn(Constants.OS.WINDOWS);
-      try (MockedStatic<SnowflakeUtil> snowflakeUtilMockedStatic =
-          Mockito.mockStatic(SnowflakeUtil.class)) {
-        snowflakeUtilMockedStatic
-            .when(() -> SnowflakeUtil.systemGetProperty("user.home"))
-            .thenReturn("/User/Home");
-        try (MockedStatic<FileUtil> fileUtilMockedStatic = Mockito.mockStatic(FileUtil.class)) {
-          fileUtilMockedStatic.when(() -> FileUtil.isWritable("/User/Home")).thenReturn(true);
-          File defaultCacheDir = FileCacheManager.getDefaultCacheDir();
-          Assertions.assertNotNull(defaultCacheDir);
-          Assertions.assertEquals(
-              "/User/Home/AppData/Local/Snowflake/Caches", defaultCacheDir.getAbsolutePath());
-        }
-      }
-    }
-  }
-
-  @Test
-  public void shouldCreateCacheDirForMacOS() {
-    try (MockedStatic<Constants> constantsMockedStatic = Mockito.mockStatic(Constants.class)) {
-      constantsMockedStatic.when(Constants::getOS).thenReturn(Constants.OS.MAC);
-      try (MockedStatic<SnowflakeUtil> snowflakeUtilMockedStatic =
-          Mockito.mockStatic(SnowflakeUtil.class)) {
-        snowflakeUtilMockedStatic
-            .when(() -> SnowflakeUtil.systemGetProperty("user.home"))
-            .thenReturn("/User/Home");
-        try (MockedStatic<FileUtil> fileUtilMockedStatic = Mockito.mockStatic(FileUtil.class)) {
-          fileUtilMockedStatic.when(() -> FileUtil.isWritable("/User/Home")).thenReturn(true);
-          File defaultCacheDir = FileCacheManager.getDefaultCacheDir();
-          Assertions.assertNotNull(defaultCacheDir);
-          Assertions.assertEquals(
-              "/User/Home/Library/Caches/Snowflake", defaultCacheDir.getAbsolutePath());
-        }
-      }
-    }
-  }
-
-  @Test
-  public void shouldReturnNullWhenNoHomeDirSet() {
-    try (MockedStatic<Constants> constantsMockedStatic = Mockito.mockStatic(Constants.class)) {
-      constantsMockedStatic.when(Constants::getOS).thenReturn(Constants.OS.LINUX);
-      try (MockedStatic<SnowflakeUtil> snowflakeUtilMockedStatic =
-          Mockito.mockStatic(SnowflakeUtil.class)) {
-        snowflakeUtilMockedStatic
-            .when(() -> SnowflakeUtil.systemGetEnv("XDG_CACHE_HOME"))
-            .thenReturn(null);
-        snowflakeUtilMockedStatic
-            .when(() -> SnowflakeUtil.systemGetProperty("user.home"))
-            .thenReturn(null);
-        File defaultCacheDir = FileCacheManager.getDefaultCacheDir();
-        Assertions.assertNull(defaultCacheDir);
-      }
-    }
   }
 
   private File createCacheFile() {
